# docs/swagger.yaml
definitions:
  entities.Priority:
    enum:
    - low
    - medium
    - high
    type: string
    x-enum-varnames:
    - Low
    - Medium
    - High
  entities.Status:
    enum:
    - pending
    - in_progress
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - Pending
    - InProgress
    - Completed
    - Cancelled
  entities.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      priority:
        $ref: '#/definitions/entities.Priority'
      status:
        $ref: '#/definitions/entities.Status'
      title:
        type: string
      updated_at:
        type: string
    type: object
  services.CreateTaskInput:
    properties:
      description:
        type: string
      due_date:
        type: string
      priority:
        $ref: '#/definitions/entities.Priority'
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - priority
    - title
    type: object
  services.UpdateTaskInput:
    properties:
      description:
        type: string
      due_date:
        type: string
      priority:
        $ref: '#/definitions/entities.Priority'
      status:
        $ref: '#/definitions/entities.Status'
      title:
        maxLength: 100
        minLength: 3
        type: string
    type: object
info:
  contact: {}
paths:
  /tasks:
    get:
      parameters:
      - description: Filter by status
        enum:
        - pending
        - in_progress
        - completed
        - cancelled
        in: query
        name: status
        type: string
      - description: Filter by priority
        enum:
        - low
        - medium
        - high
        in: query
        name: priority
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Task'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Task data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/services.CreateTaskInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a task
      tags:
      - tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Task'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get task by ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/services.UpdateTaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a task
      tags:
      - tasks
swagger: "2.0"